Continue backend debloat: split aggregate audit, Telegram, and preferences modules
Extract cohesive service and route modules behind existing facades so behavior and API contracts stay unchanged while reducing large hotspots.

Add Telegram recommendation-path tests to keep the decomposition regression-safe.
