+++ routes/preferences/constants.js
@@ -1,0 +1,15 @@
+const LASTFM_VALID_PERIODS = [
+  '7day',
+  '1month',
+  '3month',
+  '6month',
+  '12month',
+  'overall',
+];
+
+const SPOTIFY_VALID_RANGES = ['short_term', 'medium_term', 'long_term'];
+
+module.exports = {
+  LASTFM_VALID_PERIODS,
+  SPOTIFY_VALID_RANGES,
+};
