+++ services/telegram.js
@@ -2,51 +2,17 @@
 // Telegram Bot API wrapper for admin notifications
 // Supports forum topics (message_thread_id) for organized notification channels
 
-const logger = require('../utils/logger');
 const crypto = require('crypto');
-
-// Simple encryption for bot token storage
-const ENCRYPTION_ALGORITHM = 'aes-256-gcm';
-
-/**
- * Encrypt a string using AES-256-GCM
- * @param {string} text - Text to encrypt
- * @param {string} key - Encryption key (from environment)
- * @returns {string} - Encrypted string (iv:authTag:encrypted)
- */
-function encrypt(text, key) {
-  if (!key || key.length < 32) {
-    throw new Error('Encryption key must be at least 32 characters');
-  }
-  const keyBuffer = crypto.scryptSync(key, 'salt', 32);
-  const iv = crypto.randomBytes(16);
-  const cipher = crypto.createCipheriv(ENCRYPTION_ALGORITHM, keyBuffer, iv);
-  let encrypted = cipher.update(text, 'utf8', 'hex');
-  encrypted += cipher.final('hex');
-  const authTag = cipher.getAuthTag().toString('hex');
-  return `${iv.toString('hex')}:${authTag}:${encrypted}`;
-}
-
-/**
- * Decrypt a string encrypted with encrypt()
- * @param {string} encryptedText - Encrypted string
- * @param {string} key - Encryption key
- * @returns {string} - Decrypted text
- */
-function decrypt(encryptedText, key) {
-  if (!key || key.length < 32) {
-    throw new Error('Encryption key must be at least 32 characters');
-  }
-  const [ivHex, authTagHex, encrypted] = encryptedText.split(':');
-  const keyBuffer = crypto.scryptSync(key, 'salt', 32);
-  const iv = Buffer.from(ivHex, 'hex');
-  const authTag = Buffer.from(authTagHex, 'hex');
-  const decipher = crypto.createDecipheriv(ENCRYPTION_ALGORITHM, keyBuffer, iv);
-  decipher.setAuthTag(authTag);
-  let decrypted = decipher.update(encrypted, 'hex', 'utf8');
-  decrypted += decipher.final('utf8');
-  return decrypted;
-}
+const logger = require('../utils/logger');
+const { createConfigManager } = require('./telegram/config-manager');
+const { decrypt, encrypt } = require('./telegram/crypto-utils');
+const {
+  createNotificationHelpers,
+} = require('./telegram/notification-helpers');
+const {
+  createRecommendationsNotifier,
+} = require('./telegram/recommendations-notifier');
+const { createWebhookHandler } = require('./telegram/webhook-handler');
 
 // ============================================
 // SETUP HELPER FUNCTIONS
@@ -306,145 +272,6 @@ function createSetupHelpers(apiRequest, log) {
 }
 
 // ============================================
-// CONFIG MANAGER HELPER
-// ============================================
-
-/**
- * Create configuration management functions
- */
-function createConfigManager(pool, encryptionKey, log, setupHelpers, baseUrl) {
-  let configCache = null;
-  let configCacheTime = 0;
-  const CONFIG_CACHE_TTL = 60000;
-
-  async function saveConfig(config) {
-    if (!pool) throw new Error('Database pool not configured');
-
-    const encryptedToken = encrypt(config.botToken, encryptionKey);
-    const webhookSecret = crypto.randomUUID();
-
-    await pool.query('DELETE FROM telegram_config');
-
-    const result = await pool.query(
-      `INSERT INTO telegram_config 
-       (bot_token_encrypted, chat_id, thread_id, chat_title, topic_name, 
-        webhook_secret, enabled, configured_at, configured_by)
-       VALUES ($1, $2, $3, $4, $5, $6, $7, NOW(), $8)
-       RETURNING id, chat_id, thread_id, chat_title, topic_name, webhook_secret, enabled, configured_at`,
-      [
-        encryptedToken,
-        config.chatId,
-        config.threadId || null,
-        config.chatTitle,
-        config.topicName || null,
-        webhookSecret,
-        true,
-        config.configuredBy,
-      ]
-    );
-
-    const webhookResult = await setupHelpers.setWebhook(
-      config.botToken,
-      webhookSecret,
-      baseUrl
-    );
-    if (webhookResult.skipped) {
-      log.info(
-        'Webhook setup skipped - notifications will work but button callbacks will not'
-      );
-    } else if (!webhookResult.success) {
-      log.warn('Webhook setup failed:', webhookResult.error);
-    }
-
-    configCache = null;
-    const savedConfig = result.rows[0];
-    savedConfig.webhookActive = webhookResult.success;
-    return savedConfig;
-  }
-
-  async function getConfig(includeToken = false) {
-    if (!pool) return null;
-
-    if (
-      configCache &&
-      Date.now() - configCacheTime < CONFIG_CACHE_TTL &&
-      !includeToken
-    ) {
-      return configCache;
-    }
-
-    const result = await pool.query('SELECT * FROM telegram_config LIMIT 1');
-    if (result.rows.length === 0) return null;
-
-    const config = result.rows[0];
-    const sanitized = {
-      id: config.id,
-      chatId: config.chat_id,
-      threadId: config.thread_id,
-      chatTitle: config.chat_title,
-      topicName: config.topic_name,
-      webhookSecret: config.webhook_secret,
-      enabled: config.enabled,
-      recommendationsEnabled: config.recommendations_enabled || false,
-      configuredAt: config.configured_at,
-    };
-
-    if (includeToken && config.bot_token_encrypted) {
-      sanitized.botToken = decrypt(config.bot_token_encrypted, encryptionKey);
-    }
-
-    if (!includeToken) {
-      configCache = sanitized;
-      configCacheTime = Date.now();
-    }
-
-    return sanitized;
-  }
-
-  async function isConfigured() {
-    const config = await getConfig();
-    return config?.enabled || false;
-  }
-
-  async function disconnect() {
-    if (!pool) return false;
-
-    const config = await getConfig(true);
-    if (config?.botToken) {
-      try {
-        await setupHelpers.removeWebhook(config.botToken);
-      } catch (err) {
-        log.warn('Error removing webhook during disconnect:', err);
-      }
-    }
-
-    await pool.query('DELETE FROM telegram_config');
-    configCache = null;
-    log.info('Telegram disconnected');
-    return true;
-  }
-
-  async function setRecommendationsEnabled(enabled) {
-    if (!pool) return false;
-
-    await pool.query(
-      'UPDATE telegram_config SET recommendations_enabled = $1',
-      [enabled]
-    );
-    configCache = null;
-    return true;
-  }
-
-  return {
-    saveConfig,
-    getConfig,
-    isConfigured,
-    disconnect,
-    setRecommendationsEnabled,
-  };
-}
-
-// ============================================
 // MESSAGING HELPER
 // ============================================
 
@@ -522,416 +349,6 @@ function createMessenger(apiRequest, configManager, log) {
   }
 
   return { sendMessage, editMessage, answerCallbackQuery };
-}
-
-// ============================================
-// NOTIFICATION HELPER
-// ============================================
-
-/**
- * Format priority as emoji
- */
-function priorityEmoji(priority) {
-  const emojis = { urgent: '🔴', high: '🟠', normal: '🟡', low: '⚪' };
-  return emojis[priority] || '🟡';
-}
-
-/**
- * Create notification functions
- */
-function createNotificationHelpers(apiRequest, configManager, messenger, log) {
-  async function notifyNewEvent(event, actions = []) {
-    const config = await configManager.getConfig();
-    if (!config?.enabled) return null;
-
-    const emoji = priorityEmoji(event.priority);
-    let text = `${emoji} *${event.priority.toUpperCase()}* — ${event.title}\n\n`;
-
-    if (event.description) text += `${event.description}\n\n`;
-
-    if (event.data && typeof event.data === 'object') {
-      const data =
-        typeof event.data === 'string' ? JSON.parse(event.data) : event.data;
-      if (data.username) text += `👤 User: \`${data.username}\`\n`;
-      if (data.email) text += `📧 Email: \`${data.email}\`\n`;
-    }
-
-    text += `\n🆔 Event: \`${event.id.slice(0, 8)}\``;
-
-    const keyboard = [];
-    if (actions.length > 0) {
-      keyboard.push(
-        actions.map((a) => ({
-          text: a.label,
-          callback_data: `event:${event.id}:${a.id}`,
-        }))
-      );
-    }
-
-    const result = await messenger.sendMessage(text, keyboard);
-    return result.success
-      ? { messageId: result.messageId, chatId: result.chatId }
-      : null;
-  }
-
-  async function updateEventMessage(event, action, adminUsername) {
-    if (!event.telegram_message_id) return false;
-
-    const statusEmoji = { approved: '✅', rejected: '❌', dismissed: '🗑️' };
-    const emoji = statusEmoji[event.status] || '✓';
-    let text = `${emoji} *${event.status.toUpperCase()}* — ${event.title}\n\n`;
-
-    if (event.description) text += `~${event.description}~\n\n`;
-    text += `${emoji} ${action.charAt(0).toUpperCase() + action.slice(1)} by *${adminUsername}*`;
-
-    return await messenger.editMessage(event.telegram_message_id, text, []);
-  }
-
-  async function sendTestMessage() {
-    const text =
-      '✅ *SuShe Admin Notifications*\n\n' +
-      'This is a test message. Telegram notifications are working correctly!\n\n' +
-      `🕐 Sent at: ${new Date().toISOString()}`;
-    return await messenger.sendMessage(text);
-  }
-
-  async function sendTestMessageWithCredentials(token, chatId, threadId) {
-    const text =
-      '✅ *SuShe Admin Notifications*\n\n' +
-      'This is a test message. If you see this, the bot can send messages to this chat!\n\n' +
-      `🕐 Sent at: ${new Date().toISOString()}`;
-
-    try {
-      const params = { chat_id: chatId, text, parse_mode: 'Markdown' };
-      if (threadId) params.message_thread_id = threadId;
-      const result = await apiRequest(token, 'sendMessage', params);
-      return { success: true, messageId: result.message_id };
-    } catch (err) {
-      log.error('Error sending test message with credentials:', err);
-      return { success: false, error: err.message || 'Failed to send message' };
-    }
-  }
-
-  return {
-    notifyNewEvent,
-    updateEventMessage,
-    sendTestMessage,
-    sendTestMessageWithCredentials,
-  };
-}
-
-// ============================================
-// WEBHOOK HELPER
-// ============================================
-
-/**
- * Create webhook handling functions
- */
-function createWebhookHandler(pool, configManager, log) {
-  async function verifyWebhookSecret(secret) {
-    const config = await configManager.getConfig();
-    return config?.webhookSecret === secret;
-  }
-
-  function parseCallbackData(callbackData) {
-    if (!callbackData) return null;
-    const parts = callbackData.split(':');
-    if (parts[0] === 'event' && parts.length >= 3) {
-      return { type: 'event_action', eventId: parts[1], action: parts[2] };
-    }
-    return null;
-  }
-
-  async function getLinkedAdmin(telegramUserId) {
-    log.warn('[DEBUG-TELEGRAM-LINK] getLinkedAdmin called', {
-      telegramUserId,
-      telegramUserIdType: typeof telegramUserId,
-      poolExists: !!pool,
-    });
-
-    if (!pool) {
-      log.warn('[DEBUG-TELEGRAM-LINK] pool is null, returning null');
-      return null;
-    }
-
-    try {
-      const countResult = await pool.query(
-        'SELECT COUNT(*) as count FROM telegram_admins'
-      );
-      log.warn('[DEBUG-TELEGRAM-LINK] telegram_admins table count', {
-        count: countResult.rows[0]?.count,
-      });
-
-      const allAdmins = await pool.query(
-        'SELECT telegram_user_id, telegram_username, user_id FROM telegram_admins'
-      );
-      log.warn('[DEBUG-TELEGRAM-LINK] all telegram_admins entries', {
-        entries: allAdmins.rows,
-      });
-
-      const result = await pool.query(
-        `SELECT u.* FROM users u
-         JOIN telegram_admins ta ON u._id = ta.user_id
-         WHERE ta.telegram_user_id = $1`,
-        [telegramUserId]
-      );
-
-      log.warn('[DEBUG-TELEGRAM-LINK] query result', {
-        rowCount: result.rows.length,
-        foundUser: result.rows[0]?.username || null,
-      });
-
-      return result.rows[0] || null;
-    } catch (err) {
-      log.error('[DEBUG-TELEGRAM-LINK] query error', {
-        error: err.message,
-        stack: err.stack,
-      });
-      return null;
-    }
-  }
-
-  async function linkAdmin(telegramUserId, telegramUsername, appUserId) {
-    if (!pool) return false;
-
-    try {
-      await pool.query(
-        `INSERT INTO telegram_admins (telegram_user_id, telegram_username, user_id)
-         VALUES ($1, $2, $3)
-         ON CONFLICT (telegram_user_id) 
-         DO UPDATE SET telegram_username = $2, user_id = $3, linked_at = NOW()`,
-        [telegramUserId, telegramUsername, appUserId]
-      );
-      return true;
-    } catch (err) {
-      log.error('Error linking Telegram admin:', err);
-      return false;
-    }
-  }
-
-  return { verifyWebhookSecret, parseCallbackData, getLinkedAdmin, linkAdmin };
-}
-
-// ============================================
-// RECOMMENDATIONS NOTIFIER
-// ============================================
-
-/**
- * Create recommendations notification functions
- */
-function createRecommendationsNotifier(
-  pool,
-  apiRequest,
-  uploadPhoto,
-  configManager,
-  log
-) {
-  /**
-   * Get all recommendation threads from database
-   */
-  async function getThreads() {
-    if (!pool) return [];
-    const result = await pool.query(
-      'SELECT year, thread_id, topic_name, created_at FROM telegram_recommendation_threads ORDER BY year DESC'
-    );
-    return result.rows.map((row) => ({
-      year: row.year,
-      threadId: row.thread_id,
-      topicName: row.topic_name,
-      createdAt: row.created_at,
-    }));
-  }
-
-  /**
-   * Get or create a thread for a specific year
-   */
-  async function getOrCreateThread(year, botToken, chatId) {
-    if (!pool) return null;
-
-    // Check if thread exists
-    const existing = await pool.query(
-      'SELECT thread_id FROM telegram_recommendation_threads WHERE year = $1',
-      [year]
-    );
-
-    if (existing.rows.length > 0) {
-      return existing.rows[0].thread_id;
-    }
-
-    // Create new forum topic
-    const topicName = `Recommendations ${year}`;
-    try {
-      const result = await apiRequest(botToken, 'createForumTopic', {
-        chat_id: chatId,
-        name: topicName,
-      });
-
-      const threadId = result.message_thread_id;
-
-      // Save to database
-      await pool.query(
-        `INSERT INTO telegram_recommendation_threads (year, thread_id, topic_name, created_at)
-         VALUES ($1, $2, $3, NOW())`,
-        [year, threadId, topicName]
-      );
-
-      log.info('Created Telegram recommendation thread', {
-        year,
-        threadId,
-        topicName,
-      });
-      return threadId;
-    } catch (err) {
-      log.error('Failed to create recommendation thread', {
-        year,
-        error: err.message,
-      });
-      return null;
-    }
-  }
-
-  /**
-   * Send a recommendation notification to Telegram
-   * @param {Object} rec - Recommendation data
-   * @param {string} rec.artist - Artist name
-   * @param {string} rec.album - Album name
-   * @param {string} rec.album_id - Album ID for cover
-   * @param {number} rec.year - Year of recommendation
-   * @param {string} rec.recommended_by - Username of recommender
-   * @param {string} rec.reasoning - Reasoning text
-   * @param {Object} [coverImage] - Optional cover image data
-   * @param {Buffer} [coverImage.buffer] - Image buffer
-   * @param {string} [coverImage.format] - Image format (jpeg, png, etc.)
-   */
-  async function sendNotification(rec, coverImage = null) {
-    try {
-      const config = await configManager.getConfig(true);
-      if (
-        !config?.enabled ||
-        !config.recommendationsEnabled ||
-        !config.botToken
-      ) {
-        return {
-          success: false,
-          error: 'Recommendations notifications not enabled',
-        };
-      }
-
-      // Get or create thread for this year
-      const threadId = await getOrCreateThread(
-        rec.year,
-        config.botToken,
-        config.chatId
-      );
-      if (!threadId) {
-        return { success: false, error: 'Failed to get/create thread' };
-      }
-
-      // Format message
-      const message = formatRecommendationMessage(rec);
-
-      // Try to send with photo first (upload directly)
-      if (coverImage?.buffer) {
-        try {
-          await uploadPhoto(
-            config.botToken,
-            coverImage.buffer,
-            coverImage.format,
-            {
-              chat_id: config.chatId,
-              message_thread_id: threadId,
-              caption: message,
-              parse_mode: 'Markdown',
-            }
-          );
-          return { success: true };
-        } catch (photoErr) {
-          log.warn('Failed to upload photo, falling back to text', {
-            error: photoErr.message,
-          });
-        }
-      }
-
-      // Fallback to text-only message
-      await apiRequest(config.botToken, 'sendMessage', {
-        chat_id: config.chatId,
-        message_thread_id: threadId,
-        text: message,
-        parse_mode: 'Markdown',
-      });
-
-      return { success: true };
-    } catch (err) {
-      log.error('Failed to send recommendation notification', {
-        album: rec.album,
-        artist: rec.artist,
-        year: rec.year,
-        error: err.message,
-      });
-      return { success: false, error: err.message };
-    }
-  }
-
-  /**
-   * Format the recommendation message
-   */
-  function formatRecommendationMessage(rec) {
-    // Format release date as DD/MM/YYYY, fallback to year
-    let dateDisplay = rec.year;
-    if (rec.release_date) {
-      const date = new Date(rec.release_date);
-      if (!isNaN(date.getTime())) {
-        const day = String(date.getDate()).padStart(2, '0');
-        const month = String(date.getMonth() + 1).padStart(2, '0');
-        const year = date.getFullYear();
-        dateDisplay = `${day}/${month}/${year}`;
-      }
-    }
-
-    // Build RYM search URL
-    const rymSearchTerm = encodeURIComponent(`${rec.artist} ${rec.album}`);
-    const rymUrl = `https://rateyourmusic.com/search?searchterm=${rymSearchTerm}&searchtype=l`;
-
-    const lines = [
-      '🎉 *New Recommendation* 🎉',
-      '',
-      `🎵 *${escapeMarkdown(rec.album)}*`,
-      `🎤 ${escapeMarkdown(rec.artist)}`,
-      `📅 ${escapeMarkdown(dateDisplay)}`,
-      `🔗 [Rate Your Music](${rymUrl})`,
-      '',
-    ];
-
-    if (rec.reasoning) {
-      // Truncate reasoning if too long
-      const maxLen = 300;
-      let reasoning = rec.reasoning;
-      if (reasoning.length > maxLen) {
-        reasoning = reasoning.substring(0, maxLen) + '...';
-      }
-      lines.push(`💬 _"${escapeMarkdown(reasoning)}"_`);
-      lines.push('');
-    }
-
-    lines.push(`👤 Recommended by ${escapeMarkdown(rec.recommended_by)}`);
-
-    return lines.join('\n');
-  }
-
-  /**
-   * Escape special characters for Telegram Markdown v1.
-   * Only _ * ` [ need escaping in v1 (not the extended set used by MarkdownV2).
-   */
-  function escapeMarkdown(text) {
-    if (!text) return '';
-    return text
-      .replace(/_/g, '\\_')
-      .replace(/\*/g, '\\*')
-      .replace(/`/g, '\\`')
-      .replace(/\[/g, '\\[');
-  }
-
-  return { sendNotification, getThreads };
 }
 
 // ============================================
